Added Attacking hitboxes
I have added an empty game object called MeleeAttackHitboxes, this is just a holder for the others. It contains 4 other game objects. each has a 2d capsule collidor that is the hitbox for the relevent attack. Up, Down, Left and right. These are only enabled when the player attacks and are disabled after the attack is completed. The hit boxes for attacking are a bit big for now but can be changed.