Added a Spell Attack
Hold Left Shift Then left click to perform a lighning spell attack.
Added a playerHolder gameobject (this is empty and just to hold the player objects)
Inside it has the player and a new object called spellAttack that runs the spell animation and animator.
Updated the player controller script to add the spell attack.