Added Credits and FadeIn/FadeOut
added a Credits Scene.
You can access it from the main menu.
you can hold Space to skip the credits.
It automatically goes back to the main menu once the credits finish.

Added fades.
It now fades out from the main menu when you change scenes. It fades in to the credits and fades out from the credits.