Added IFrame Animation
The sprite will now flicker when the player takes damage. This isnt done through animation but sounded best for the update summary. I have added code to the health bar.cs so when the player takes damage it uses InvokeRepeating to call a function which enables and disenables the sprite renderer.