Fixed bug & changed other things
Fixed a bug where the player could move the outer walls.
To fix it i changed the walls rigid body to static so they they cant be moved.

I also tidied the folders a bit and moved some scripts i made into the correct folders

Also edited the players hitbox so that it colliders with the walls better

Removed old script i made that wasnt needed

Moved the wall hitbox to be more accurate.