Fixes and implemented basic death screen
added a death screen where the user can press R to restart.

Fixed a bunch of errors that was caused from deleting the player game object upon death. Now we just disable the player sprite renderer and set timescale to 0.

This is still placeholder until a death animation is created, after that, we will play the death animation, freeze it on the last frame, then the death menu will appear and nothing will be deleted or disabled.